+++ CHANGELOG.md
@@ -46,3 +46,5 @@
 * Fixed a bug where the number thresholds had to be exact instead of between two thresholds.
 * Fixed a bug where only the first word in the tooltip key was capitalized instead of the key being in title case.
 * Fixed the regex for numeric comparisions so that it would take negative numbers into account.
+* Added additional tests.
+* Fixed the way that the width of the parent labels was determined and added a cache.